+++ src/constant/walletType.js
@@ -16,6 +16,7 @@ export const ACCOUNT_TYPE = {
 export const NET_CONFIG_TYPE = {
   Mainnet:"Mainnet",
   Devnet:"Devnet",
+  QAnet:"Berkeley-QA",
   Unknown:"Unknown",
 }
 
@@ -24,3 +25,15 @@ export const WALLET_CONNECT_TYPE = {
   WALLET_APP_CONNECT:"WALLET_APP_CONNECT",
   CONTENT_SCRIPT : 'mina-contentscript',
 }
+
+
+
+/** not support transaction history */
+export const NET_CONFIG_NOT_SUPPORT_TX_HISTORY =[
+  NET_CONFIG_TYPE.QAnet,
+  NET_CONFIG_TYPE.Unknown,
+]
+/** not support stake   */
+export const NET_CONFIG_NOT_SUPPORT_STAKING =[
+  NET_CONFIG_TYPE.Unknown,
+]